requirements.txt
PyQt6>=6.0
pywebview>=3.0
PyPDF2>=3.0  # optional (PDF extraction support)